@@ -1,10 +1,10 @@
 name: fit_kit
 description: Flutter plugin for reading health and fitness data. Wraps HealthKit on iOS and GoogleFit on Android.
-version: 1.1.2
+version: 1.1.3
 homepage: https://github.com/krokyze/fitkit
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
