@@ -5,7 +5,7 @@ class FitKit {
 
   /// iOS isn't completely supported by HealthKit, false means no, true means user has approved or declined permissions.
   /// In case user has declined permissions read will just return empty list for declined data types.
-  static Future<bool> hasPermissions(List<DataType> types) async {
+  static Future<bool?> hasPermissions(List<DataType> types) async {
     return await _channel.invokeMethod('hasPermissions', {
       "types": types.map((type) => _dataTypeToString(type)).toList(),
     });
@@ -15,7 +15,7 @@ class FitKit {
   /// otherwise iOS HealthKit will ask to approve every permission one by one in separate screens.
   ///
   /// `await FitKit.requestPermissions(DataType.values)`
-  static Future<bool> requestPermissions(List<DataType> types) async {
+  static Future<bool?> requestPermissions(List<DataType> types) async {
     return await _channel.invokeMethod('requestPermissions', {
       "types": types.map((type) => _dataTypeToString(type)).toList(),
     });
@@ -29,9 +29,9 @@ class FitKit {
   /// #### It's not advised to call `await FitKit.read(dataType)` without any extra parameters. This can lead to FAILED BINDER TRANSACTION on Android devices because of the data batch size being too large.
   static Future<List<FitData>> read(
     DataType type, {
-    DateTime dateFrom,
-    DateTime dateTo,
-    int limit,
+    DateTime? dateFrom,
+    DateTime? dateTo,
+    int? limit,
   }) async {
     return await _channel
         .invokeListMethod('read', {
@@ -41,7 +41,8 @@ class FitKit {
           "limit": limit,
         })
         .then(
-          (response) => response.map((item) => FitData.fromJson(item)).toList(),
+          (response) =>
+              response!.map((item) => FitData.fromJson(item)).toList(),
         )
         .catchError(
           (_) => throw UnsupportedException(type),
@@ -50,11 +51,6 @@ class FitKit {
             return false;
           },
         );
-  }
-
-  static Future<FitData> readLast(DataType type) async {
-    return await read(type, limit: 1)
-        .then((results) => results.isEmpty ? null : results[0]);
   }
 
   static String _dataTypeToString(DataType type) {
